Tracer advection utest I (#463)
* modifies two advection stencils (mo_advection_btraj_compute). Theses changes were needed for their integration in the blue-line.
* adds the advection tests in the tox tests.

---------

+++ model/atmosphere/advection/tests/advection_stencil_tests/test_prep_gauss_quadrature_c_list_stencil.py
@@ -99,7 +99,7 @@ class TestPrepGaussQuadratureCListStencil(StencilTest):
             1.0 + zeta_4,
         )
 
-        famask_bool = np.where(famask_int == np.int32(1), True, False)
+        famask_bool = np.where(famask_int == 1, True, False)
 
         p_coords_dreg_v_1_x = np.where(famask_bool, p_coords_dreg_v_1_x, 0.0)
         p_coords_dreg_v_2_x = np.where(famask_bool, p_coords_dreg_v_2_x, 0.0)
