+++ model/atmosphere/diffusion/tests/diffusion_stencil_tests/test_temporary_fields_for_turbulence_diagnostics.py
@@ -73,8 +73,8 @@ class TestTemporaryFieldsForTurbulenceDiagnostics(StencilTest):
             diff_multfac_smag=diff_multfac_smag,
             kh_c=kh_c,
             div=div,
-            horizontal_start=int32(0),
+            horizontal_start=0,
             horizontal_end=int32(grid.num_cells),
-            vertical_start=int32(0),
+            vertical_start=0,
             vertical_end=int32(grid.num_levels),
         )
